Business e Detalhe do Imóvel

+++ Imob_App/imob_app.business/Contratos/IEntidadeLocalizacao.cs
@@ -11,6 +11,6 @@ namespace imob_app.business.Contratos
 {
     public interface IEntidadeLocalizacao<T> : IEntidade<T>
     {
-        List<T> SelecionarPorReferencia(short id);
+        List<T> SelecionarPorReferencia(int id);
     }
 }
